+++ example/targets/live-activities/LiveActivityExampleBundle.swift
@@ -1,0 +1,11 @@
+import SwiftUI
+import WidgetKit
+
+#if canImport(ActivityKit)
+  @main
+  struct LiveActivityExampleBundle: WidgetBundle {
+    var body: some Widget {
+      LiveActivityExample()
+    }
+  }
+#endif